+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/theming/ThemeController.kt
@@ -1,0 +1,11 @@
+package com.altmann.choresmanager.theming
+
+import androidx.compose.runtime.MutableState
+
+class ThemeController(
+    val themeState: MutableState<AppTheme>
+) {
+    fun setPrimary(appTheme: AppTheme) {
+        themeState.value = appTheme
+    }
+}